requirements.txt
accelerate
fastapi
pyserial
python_dotenv
torch
transformers
uvicorn